implement template based display of UI results - first pass (#8)
