+++ plugin/MetaSearch/resources/templates/record_metadata_dc.html
@@ -1,0 +1,84 @@
+<!DOCTYPE html>
+<html lang="{{ language }}">
+    <head>
+        <meta charset="utf-8"/>
+        <title>{{ labels['record-metadata-dc'] }}</title>
+        <style type="text/css">
+            body {
+                background-color: #ffffff;
+                font-family: arial, verdana, sans-serif;
+                text-align: left;
+                float: left;
+            }
+            header {
+                display: inline-block;
+            }
+        </style>
+    </head>
+    <body>
+        <header>
+            <h3>{{ labels['record-metadata-dc'] }}</h3>
+        </header>
+        <section id="record-metadata">
+            <table>
+                <tr>
+                    <th>{{ labels['property'] }}</th>
+                    <th>{{ labels['value'] }}</th>
+                </tr>
+                <tr>
+                    <td>{{ labels['identifier'] }}</td>
+                    <td>{{ obj.identifier }}</td>
+                </tr>
+                <tr>
+                    <th>{{ labels['title'] }}</th>
+                    <th>{{ obj.title }}</th>
+                </tr>
+                <tr>
+                    <td>{{ labels['abstract'] }}</td>
+                    <td>{{ obj.abstract }}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['subjects'] }}</td>
+                    <td>{{ obj.subjects|join(',') }}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['creator'] }}</td>
+                    <td>{{ obj.creator }}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['contributor'] }}</td>
+                    <td>{{ obj.contributor}}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['publisher'] }}</td>
+                    <td>{{ obj.publisher}}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['contributor'] }}</td>
+                    <td>{{ obj.contributor}}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['modified'] }}</td>
+                    <td>{{ obj.modified }}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['language'] }}</td>
+                    <td>{{ obj.language }}</td>
+                </tr>
+
+                <tr>
+                    <td>{{ labels['format'] }}</td>
+                    <td>{{ obj.format }}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['rights'] }}</td>
+                    <td>{{ obj.rights|join(',') }}</td>
+                </tr>
+                <tr>
+                    <td>{{ labels['bbox'] }}</td>
+                    <td>{{ [obj.bbox.minx, obj.bbox.miny, obj.bbox.maxx, obj.bbox.maxy]|join(',') }}</td>
+                </tr>
+            </table>"
+        </section>
+    </body>
+</html>
