+++ requirements.txt
@@ -1,2 +1,3 @@
+Jinja2==2.7.2
 OWSLib==0.8.4
 pygments==1.6
